Improved logging, randomized group id to improve reliability.
https://github.com/edenhill/librdkafka/issues/3226#issuecomment-765560247

+++ tests/Foundatio.Kafka.Tests/Messaging/KafkaMessageBusTestBase.cs
@@ -2,25 +2,26 @@
 using System.Threading.Tasks;
 using Foundatio.Messaging;
 using Foundatio.Tests.Messaging;
-using Foundatio.Utility;
 using Xunit.Abstractions;
 
 namespace Foundatio.Kafka.Tests.Messaging;
 
 public class KafkaMessageBusTestBase : MessageBusTestBase {
-    protected readonly string _topic = $"test_{SystemClock.UtcNow.Ticks}";
+    protected readonly string Topic = $"test_{Guid.NewGuid():N}";
+    protected readonly string GroupId = $"group_{Guid.NewGuid():N}";
 
     public KafkaMessageBusTestBase(ITestOutputHelper output) : base(output)
     {
+        EnableTopicDeletion = true;
     }
 
     protected override IMessageBus GetMessageBus(Func<SharedMessageBusOptions, SharedMessageBusOptions> config = null) {
         return new KafkaMessageBus(o => o
             .BootstrapServers("127.0.0.1:9092")
-            .Topic(_topic)
+            .Topic(Topic)
             .TopicReplicationFactor(1)
             .TopicNumberOfPartitions(1)
-            .GroupId("tests")
+            .GroupId(GroupId)
             .EnableAutoCommit(false)
             .EnableAutoOffsetStore(false)
             .AllowAutoCreateTopics(true)
@@ -31,7 +32,9 @@ public class KafkaMessageBusTestBase : MessageBusTestBase {
     protected override async Task CleanupMessageBusAsync(IMessageBus messageBus) {
         await base.CleanupMessageBusAsync(messageBus);
 
-        if (messageBus is IKafkaMessageBus kafkaMessageBus)
+        if (EnableTopicDeletion && messageBus is IKafkaMessageBus kafkaMessageBus)
             await kafkaMessageBus.DeleteTopicAsync();
     }
+
+    public bool EnableTopicDeletion { get; set; }
 }
